use double to represent lat/lon

+++ TaxiData/src/main/java/com/gsoeller/taxi/pojos/Location.java
@@ -1,22 +1,22 @@
 package com.gsoeller.taxi.pojos;
 
 public class Location {
-	private long latitude;
-	private long longitude;
+	private double latitude;
+	private double longitude;
 
-	public long getLongitude() {
+	public double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(long longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
 
-	public long getLatitude() {
+	public double getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(long latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
